pyproject.toml
[project]
name = "debezium-db-sync"
version = "0.1.0"
description = ""
authors = [
    {name = "shahnoor77",email = "shahnoorkhan9955@gmail.com"}
]

readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "mysql-connector-python (>=9.5.0,<10.0.0)",
    "jpype1 (>=1.6.0,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[tool.poetry]
packages = [
    { include = "db_sync", from = "src" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
